Modify: activities module logic

+++ apps/server/src/activities/activities.controller.ts
@@ -7,13 +7,19 @@ import {
   Param,
   ParseIntPipe,
   Delete,
+  Req,
+  UnauthorizedException,
+  UseGuards,
 } from '@nestjs/common';
 import { ActivitiesService } from './activities.service';
 import { CreateActivityDto } from './dto/create-activity.dto';
 import { UpdateActivityDto } from './dto/update-user.dto';
 import { ok } from 'src/utils/response.util';
-import { ApiCreatedResponse, ApiTags } from '@nestjs/swagger';
+import { ApiBearerAuth, ApiCreatedResponse, ApiTags } from '@nestjs/swagger';
+import JWTGuard from 'src/auth/guards/jwt.guard';
 
+@UseGuards(JWTGuard)
+@ApiBearerAuth()
 @ApiTags('Activities')
 @Controller('activities')
 export class ActivitiesController {
@@ -35,14 +41,21 @@ export class ActivitiesController {
   @Patch(':id')
   async update(
     @Param('id', ParseIntPipe) id: number,
+    @Req() req: any,
     @Body() updateActivityDto: UpdateActivityDto,
   ) {
+    if (!this.activitiesService.checkIfCreator(req.user.sub, id))
+      throw new UnauthorizedException('You are not the creator');
+
     const data = await this.activitiesService.update(id, updateActivityDto);
     return ok('Updated activity successfully', data);
   }
 
   @Delete(':id')
-  async delete(@Param('id', ParseIntPipe) id: number) {
+  async delete(@Param('id', ParseIntPipe) id: number, @Req() req: any) {
+    if (!this.activitiesService.checkIfCreator(req.user.sub, id))
+      throw new UnauthorizedException('You are not the creator');
+
     const data = await this.activitiesService.delete(id);
     return ok('Deleted activity successfully', data);
   }
